+++ rpcs3/Emu/RSX/RSXThread.h
@@ -6,6 +6,7 @@
 #include "rsx_cache.h"
 #include "RSXFIFO.h"
 #include "RSXTexture.h"
+#include "RSXOffload.h"
 #include "RSXVertexProgram.h"
 #include "RSXFragmentProgram.h"
 #include "rsx_methods.h"
@@ -296,62 +297,6 @@ namespace rsx
 			return mem;
 		}
 	};
-
-	class dma_manager
-	{
-		enum op
-		{
-			raw_copy = 0,
-			vector_copy = 1,
-			index_emulate = 2
-		};
-
-		struct transport_packet
-		{
-			op type;
-			std::vector<u8> opt_storage;
-			void *src;
-			void *dst;
-			u32 length;
-			u32 aux_param0;
-			u32 aux_param1;
-
-			transport_packet(void *_dst, void *_src, u32 len)
-				: src(_src), dst(_dst), length(len), type(op::raw_copy)
-			{}
-
-			transport_packet(void *_dst, std::vector<u8>& _src, u32 len)
-				: dst(_dst), opt_storage(std::move(_src)), length(len), type(op::vector_copy)
-			{}
-
-			transport_packet(void *_dst, rsx::primitive_type prim, u32 len)
-				: dst(_dst), aux_param0(static_cast<u8>(prim)), length(len), type(op::index_emulate)
-			{}
-		};
-
-		std::deque<transport_packet> m_work_queue;
-		std::thread m_worker_thread;
-		std::mutex  m_queue_mutex;
-		thread_state m_worker_state;
-
-	public:
-		dma_manager() = default;
-
-		// initialization
-		void init();
-
-		// General tranport
-		void copy(void *dst, std::vector<u8>& src, u32 length);
-		void copy(void *dst, void *src, u32 length);
-
-		// Vertex utilities
-		void emulate_as_indexed(void *dst, rsx::primitive_type primitive, u32 count);
-
-		// Synchronization
-		void sync();
-		void join();
-	};
-	extern dma_manager g_dma_manager;
 
 	struct framebuffer_layout
 	{
